refactor: rename use case package
