+++ cmd/game/start.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"elton-okawa/battleship/internal/use_case/game"
+	"elton-okawa/battleship/internal/usecase/game"
 )
 
 type Start struct {
