+++ internal/usecase/game/interfaces.go
