+++ internal/usecase/ucerror/error_code.go
@@ -1,4 +1,4 @@
-package errors
+package ucerror
 
 const (
 	GenericError    = 0 // errors which user does not have any action
